requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-allauth>=0.57,<1.0
django-cors-headers>=4.3,<5.0
Pillow>=10.0,<11.0
boto3>=1.28,<2.0
azure-storage-blob>=12.19,<13.0
google-cloud-storage>=2.14,<3.0
paramiko>=3.3,<4.0
python-dotenv>=1.0,<2.0
cryptography>=41.0,<42.0
psycopg2-binary>=2.9,<3.0
gunicorn>=21.2,<22.0
